# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Makers Community API",
        "description": "API complète pour la plateforme communautaire Makers Community avec authentification, posts, commentaires, ratings et modération automatique. Cette API permet de gérer une communauté de makers avec toutes les fonctionnalités nécessaires.",
        "contact": {
            "name": "Équipe Makers Community",
            "email": "dev@makerscomm.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://api.makerscomm.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/posts/{post}/reactions": {
            "get": {
                "tags": [
                    "Reactions"
                ],
                "summary": "Obtenir les statistiques de réactions d'un post",
                "operationId": "58a4d13668bb3a83334768ab6f44acf3",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID du post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistiques récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "post_id": {
                                            "type": "integer"
                                        },
                                        "likes_count": {
                                            "type": "integer"
                                        },
                                        "dislikes_count": {
                                            "type": "integer"
                                        },
                                        "user_reaction": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reactions"
                ],
                "summary": "Ajouter ou modifier une réaction sur un post",
                "operationId": "026d8aa251aeab36615847f9024c5961",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID du post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "like",
                                            "dislike"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réaction ajoutée ou modifiée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reaction": {
                                            "type": "object",
                                            "nullable": true
                                        },
                                        "likes_count": {
                                            "type": "integer"
                                        },
                                        "dislikes_count": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reactions"
                ],
                "summary": "Supprimer la réaction de l'utilisateur sur un post",
                "operationId": "1ab38d9bd0e4d44b0854bc901c6c53c1",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID du post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réaction supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/posts/{post}/reactions/user/me": {
            "get": {
                "tags": [
                    "Reactions"
                ],
                "summary": "Obtenir la réaction de l'utilisateur connecté sur un post",
                "operationId": "65deb11bd0ec3e295120a055db3a6272",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID du post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réaction trouvée ou aucune réaction",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reaction": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token d'authentification Laravel Sanctum. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}